Add Technical Requirements section to README
This commit adds a Technical Requirements section to the tasks README file. Due
to recent feedback from candidates, we failed to disclose what version of `node`
and `yarn` should be used to start and build the task.

By adding this section, it should be clear to candidates what they need to do if
they run into any problems building the task.

Future work will be done to update the supported Node version and update any
dependencies that depend on it.