Update tooling for Calendar Task
This commit updates the front-end specific tooling for the Calendar task and
does the following.

*React Scripts*
Updates `react-scripts` to [`4.0.3`]
(https://github.com/facebook/create-react-app/releases/tag/v4.0.3) to ensure
that we are using the latest tools avaliable for Create React App.

*Node Sass*
This commit drops the dependency of `node-sass` which was causing issues when
candidates are using newer versions of Node which weren't ~v14.0.0. Also,
`node-sass` is [now end of life and new features are not currently being added to
it.](https://github.com/sass/node-sass#node-sass)

Instead, this commit installs `sass` (Dart Sass) to the repository which is the
primary implementation of Sass and is reccommended by [Create React App](https://create-react-app.dev/docs/adding-a-sass-stylesheet/)